% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math_functions.R
\name{multiply_numbers}
\alias{multiply_numbers}
\title{Multiply two numbers.}
\usage{
multiply_numbers(a, b)
}
\arguments{
\item{a}{A number.}

\item{b}{A number.}
}
\value{
The product of a and b.
}
\description{
Multiply two numbers.
}
\examples{
multiply_numbers(2, 2)

}
